+++ NEMODU/NEMODU/Global/Network/APISession.swift
@@ -11,74 +11,6 @@ import Foundation
 
 struct APISession: APIService {
     
-    /// [POST] 헤더에 kakaoAccessToken을 붙여 로그인을 요청하는 메서드
-    func kakaoLoginRequest<T: Decodable>(with urlResource: urlResource<T>, param: Parameters) -> Observable<Result<T, APIError>> {
-        Observable<Result<T, APIError>>.create { observer in
-            guard let kakaoAccessToken = UserDefaults.standard.string(forKey: UserDefaults.Keys.kakaoAccessToken) else { fatalError() }
-            
-            let headers: HTTPHeaders = [
-                "Content-Type": "application/json",
-                "Kakao-Access-Token": kakaoAccessToken
-            ]
-            
-            let task = AF.request(urlResource.resultURL,
-                                  method: .post,
-                                  parameters: param,
-                                  encoding: JSONEncoding.default,
-                                  headers: headers)
-                .validate(statusCode: 200...399)
-                .responseDecodable(of: T.self) { response in
-                    switch response.result {
-                    case .failure:
-                        observer.onNext(urlResource.judgeError(statusCode: response.response?.statusCode ?? -1))
-                        
-                    case .success(let data):
-                        observer.onNext(.success(data))
-                        guard let headers = response.response?.headers,
-                              let authToken = headers.value(for: "Authorization"),
-                              let refreshToken = headers.value(for: "Refresh-Token") else { return }
-                        UserDefaults.standard.set(authToken, forKey: UserDefaults.Keys.accessToken)
-                        UserDefaults.standard.set(refreshToken, forKey: UserDefaults.Keys.refreshToken)
-                    }
-                }
-            
-            return Disposables.create {
-                task.cancel()
-            }
-        }
-    }
-    
-    /// [GET] 헤더에 kakaoAccessToken을 붙여 기존 유저인지 확인을 요청하는 메서드
-    func checkOriginUser<T>(with urlResource: urlResource<T>) -> Observable<Result<T, APIError>> where T : Decodable {
-        
-        Observable<Result<T, APIError>>.create { observer in
-            guard let kakaoAccessToken = UserDefaults.standard.string(forKey: UserDefaults.Keys.kakaoAccessToken) else { fatalError() }
-            
-            let headers: HTTPHeaders = [
-                "Content-Type": "application/json",
-                "Access-Token": kakaoAccessToken
-            ]
-            
-            let task = AF.request(urlResource.resultURL,
-                                  encoding: URLEncoding.default,
-                                  headers: headers)
-                .validate(statusCode: 200...399)
-                .responseDecodable(of: T.self) { response in
-                    switch response.result {
-                    case .failure:
-                        observer.onNext(urlResource.judgeError(statusCode: response.response?.statusCode ?? -1))
-                        
-                    case .success(let data):
-                        observer.onNext(.success(data))
-                    }
-                }
-            
-            return Disposables.create {
-                task.cancel()
-            }
-        }
-    }
-    
     /// [GET]
     func getRequest<T>(with urlResource: urlResource<T>) -> Observable<Result<T, APIError>> where T : Decodable {
